refactor(core): extract content-control property helpers and complete the field-link test scheme check
Move the pure SDT property helpers (wmlValOf, contentControlPropertiesOf,
propertyChild, propertyVal, parseContentControlLock, mapContentControlType,
controlLevelOf) out of semantic-layout.ts into a new
content-control-properties.ts sibling, dropping semantic-layout.ts back under
the max-lines cap. Broaden the field-link projector test double to refuse the
full dangerous-scheme set (javascript/data/vbscript/file), matching the real
sanitizer and clearing the CodeQL incomplete-URL-scheme alert.
